Event viewer framework
Setup event proccessing for a single packet and layed out the framework
for others. Note that it only works in debug mode for now because it
somehow forces the subprocess to run in unbuffered mode even though the
flag is passed (-u) to it in the subprocess command. It doesn't work
with the regular cli call. Going to have to impliment `flush=True` or
`sys.stdout.flush()` when working on https://github.com/RMIT-Competition-Rocketry/GCS/issues/5