Pendant starts (testing needed, mac broken)
My mac thinks I'm running malware. I added python and terminal to the
allowed list in accessability settings but still no luck. Restarted
laptop and no luck. Run with sudo, no luck. It comes up with the lock,
but I can't type in. Even if i disable the echo settings. Idk. It's too
much terminal environment fuckery to get it to work. Just commiting this
to the remote so I can load it onto the pi and give it a go.